DocChat - AI/docchatAI-main/docchatAI-main/local_qa_app/requirements.txt
streamlit>=1.31.0
torch>=2.2.0
transformers>=4.37.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
accelerate>=0.27.0
bitsandbytes>=0.41.0  # For 4-bit quantization
safetensors>=0.4.0 